requirements.txt
# Core dependencies
pymupdf>=1.24.0
google-genai>=0.8.0
pydantic>=2.0.0
jinja2>=3.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
playwright>=1.40.0.0
pillow>=10.0.0
pytest>=8.0.0

# Optional — PDF export (requires GTK runtime on Windows)
# Install GTK first: https://github.com/tschoonj/GTK-for-Windows-Runtime-Environment-Installer
# weasyprint>=61.0

